refactor(ui): centralize time range constants to apply DRY principle
- Create _constants/time-range.constants.ts as single source of truth
- Remove duplicated TIME_RANGE_OPTIONS, TIME_RANGE_DAYS, DEFAULT_DAYS
- SSR and client component now use shared DEFAULT_TIME_RANGE
- Action uses shared getDateFromForTimeRange function
- Reduces code duplication and ensures consistency
