+++ ui/app/(prowler)/_new-overview/severity-over-time/_constants/index.ts
@@ -1,0 +1,1 @@
+export * from "./time-range.constants";
